fix: move isEqual to util, add null check
